requirements.txt
yfinance
pandas
yahoo_fin
google-auth-oauthlib
google-auth
requests

